pat, variable ambiguous. how to fix?
